Implement sending health parameters to back-end

+++ app/src/main/java/by/iapsit/healthandlife/domain/model/User.kt
@@ -15,5 +15,5 @@ data class User(
     val email: String,
     val heartRate: Int,
     val saturation: Int,
-    val temperature: Int
+    val temperature: Double
 )
